Refactor event broadcasting and integrate setup function
- Removed the old broadcastEvent helper function and replaced it with a setupBroadcastEvent function for improved organization.
- Updated the ping route to utilize the global broadcastEvent function for sending events to specific users or all clients.
- Enhanced the structure of the event handling to streamline the broadcasting process.
