+++ app/api/events/route.ts
@@ -2,6 +2,7 @@ import { NextRequest } from "next/server";
 import { auth } from "@clerk/nextjs/server";
 import { SESSION_ID_KEY } from "@/lib/session";
 import { EventData } from "@/lib/events";
+import { setupBroadcastEvent } from "@/lib/events/broadcast";
 
 // Track connection attempts for debugging
 const connectionAttempts = new Map();
@@ -151,63 +152,6 @@ export async function GET(request: NextRequest) {
   return new Response(stream, { headers });
 }
 
-// Helper function to broadcast events to specific clients or all clients
-export function broadcastEvent(
-  event: string,
-  data: EventData,
-  targetIds?: string[]
-) {
-  if (!global.sseClients || global.sseClients.length === 0) {
-    console.log("No SSE clients to broadcast to");
-    return;
-  }
-
-  console.log(
-    `Broadcasting event ${event} to ${
-      targetIds ? targetIds.length + " specific" : "all"
-    } clients`
-  );
-
-  // Check if we're targeting specific clients
-  if (targetIds && targetIds.length > 0) {
-    // Find the target clients
-    const targetClients = global.sseClients.filter((client) =>
-      targetIds.includes(client.id)
-    );
-
-    if (targetClients.length === 0) {
-      console.log(
-        `Warning: No matching clients found for target IDs: ${targetIds.join(
-          ", "
-        )}`
-      );
-      return;
-    }
-
-    console.log(`Found ${targetClients.length} matching clients for broadcast`);
-  }
-
-  let sentCount = 0;
-  global.sseClients.forEach((client) => {
-    // Send to specific clients if targetIds is provided, otherwise send to all
-    if (!targetIds || targetIds.includes(client.id)) {
-      try {
-        console.log(`Sending ${event} event to client ${client.id}`);
-        client.send({
-          type: event,
-          data,
-          timestamp: new Date().toISOString(),
-        });
-        sentCount++;
-      } catch (error) {
-        console.error(`Error sending event to client ${client.id}:`, error);
-      }
-    }
-  });
-
-  console.log(`Successfully sent ${event} event to ${sentCount} clients`);
-}
-
 // Clean up dead connections periodically
 const CLEANUP_INTERVAL = 5 * 60 * 1000; // 5 minutes
 const CONNECTION_TIMEOUT = 10 * 60 * 1000; // 10 minutes
@@ -248,6 +192,9 @@ if (typeof global !== "undefined") {
     console.log("SSE connection cleanup interval initialized");
   }
 }
+
+// Setup the broadcast event function
+setupBroadcastEvent();
 
 // Tell Next.js this is a dynamic API route that shouldn't be cached
 export const dynamic = "force-dynamic";
