+++ lib/events/broadcast.ts
@@ -1,0 +1,78 @@
+import { EventData } from "@/lib/events";
+
+// Helper function to broadcast events to specific clients or all clients
+export function broadcastEvent(
+  event: string,
+  data: EventData,
+  targetIds?: string[]
+) {
+  if (!global.sseClients || global.sseClients.length === 0) {
+    console.log("No SSE clients to broadcast to");
+    return;
+  }
+
+  console.log(
+    `Broadcasting event ${event} to ${
+      targetIds ? targetIds.length + " specific" : "all"
+    } clients`
+  );
+
+  // Check if we're targeting specific clients
+  if (targetIds && targetIds.length > 0) {
+    // Find the target clients
+    const targetClients = global.sseClients.filter((client) =>
+      targetIds.includes(client.id)
+    );
+
+    if (targetClients.length === 0) {
+      console.log(
+        `Warning: No matching clients found for target IDs: ${targetIds.join(
+          ", "
+        )}`
+      );
+      return;
+    }
+
+    console.log(`Found ${targetClients.length} matching clients for broadcast`);
+  }
+
+  let sentCount = 0;
+  global.sseClients.forEach((client) => {
+    // Send to specific clients if targetIds is provided, otherwise send to all
+    if (!targetIds || targetIds.includes(client.id)) {
+      try {
+        console.log(`Sending ${event} event to client ${client.id}`);
+        client.send({
+          type: event,
+          data,
+          timestamp: new Date().toISOString(),
+        });
+        sentCount++;
+      } catch (error) {
+        console.error(`Error sending event to client ${client.id}:`, error);
+      }
+    }
+  });
+
+  console.log(`Successfully sent ${event} event to ${sentCount} clients`);
+}
+
+// Function to set up the broadcast event - to be called from the route.ts file
+export function setupBroadcastEvent() {
+  // Make broadcastEvent available globally
+  if (typeof global !== "undefined") {
+    global.broadcastEvent = broadcastEvent;
+  }
+}
+
+// Add TypeScript declaration
+type BroadcastEventFn = (
+  event: string,
+  data: EventData,
+  targetIds?: string[]
+) => void;
+
+declare global {
+  // eslint-disable-next-line no-var
+  var broadcastEvent: BroadcastEventFn;
+}
